projects/telegram-bot/requirements.txt
py-algorand-sdk==2.6.1
python-dotenv==1.0.1
python-telegram-bot==21.6
